+++ src/Snappy.Standard/SnappyStatus.cs
@@ -1,0 +1,9 @@
+namespace Snappy
+{
+    enum SnappyStatus
+    {
+        Ok = 0,
+        InvalidInput = 1,
+        BufferTooSmall = 2
+    }
+}