+++ tests/AvroConvertTests/CodecTests/TwitterExampleTests.cs
@@ -15,8 +15,6 @@ namespace AvroConvertTests.CodecTests
             _snappy = System.IO.File.ReadAllBytes("CodecTests/twitter.snappy.avro");
         }
 
-        //TODO ComponentTestForSnappy
-
         [Fact]
         public void Deserialize_NotCompressed_DataIsDeserialized()
         {
@@ -35,7 +33,6 @@ namespace AvroConvertTests.CodecTests
                 Tweet = "Works as intended.  Terran is IMBA.",
                 Username = "BlizzardCS"
             });
-
 
             //Act
             var result = AvroConvert.Deserialize<List<TwitterModel>>(_notCompressed);